refactor: break the "halo2_frontend/src/plonk/circuit/" files
