+++ halo2_frontend/src/plonk/circuit/expression.rs
@@ -1,443 +1,13 @@
-use crate::circuit::Region;
-use crate::plonk::circuit::{Advice, ColumnType, Fixed, Instance, VirtualCells};
-use crate::plonk::Error;
+use crate::plonk::circuit::{
+    Advice, AdviceQuery, Challenge, Column, Fixed, FixedQuery, Instance, InstanceQuery, Selector,
+    VirtualCells,
+};
 use core::cmp::max;
 use core::ops::{Add, Mul};
-use halo2_middleware::circuit::{Any, ChallengeMid, ColumnMid, ExpressionMid, QueryMid, VarMid};
+use halo2_middleware::circuit::{Any, ExpressionMid, QueryMid, VarMid};
 use halo2_middleware::ff::Field;
-use halo2_middleware::poly::Rotation;
-use sealed::SealedPhase;
-use std::fmt::Debug;
 use std::iter::{Product, Sum};
-use std::{
-    convert::TryFrom,
-    ops::{Neg, Sub},
-};
-
-/// A column with an index and type
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub struct Column<C: ColumnType> {
-    pub index: usize,
-    pub column_type: C,
-}
-
-impl From<Column<Any>> for ColumnMid {
-    fn from(val: Column<Any>) -> Self {
-        ColumnMid {
-            index: val.index(),
-            column_type: (*val.column_type()),
-        }
-    }
-}
-
-impl<C: ColumnType> Column<C> {
-    pub fn new(index: usize, column_type: C) -> Self {
-        Column { index, column_type }
-    }
-
-    /// Index of this column.
-    pub fn index(&self) -> usize {
-        self.index
-    }
-
-    /// Type of this column.
-    pub fn column_type(&self) -> &C {
-        &self.column_type
-    }
-
-    /// Return expression from column at a relative position
-    pub fn query_cell<F: Field>(&self, at: Rotation) -> Expression<F> {
-        self.column_type.query_cell(self.index, at)
-    }
-
-    /// Return expression from column at the current row
-    pub fn cur<F: Field>(&self) -> Expression<F> {
-        self.query_cell(Rotation::cur())
-    }
-
-    /// Return expression from column at the next row
-    pub fn next<F: Field>(&self) -> Expression<F> {
-        self.query_cell(Rotation::next())
-    }
-
-    /// Return expression from column at the previous row
-    pub fn prev<F: Field>(&self) -> Expression<F> {
-        self.query_cell(Rotation::prev())
-    }
-
-    /// Return expression from column at the specified rotation
-    pub fn rot<F: Field>(&self, rotation: i32) -> Expression<F> {
-        self.query_cell(Rotation(rotation))
-    }
-}
-
-impl<C: ColumnType> Ord for Column<C> {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        // This ordering is consensus-critical! The layouters rely on deterministic column
-        // orderings.
-        match self.column_type.into().cmp(&other.column_type.into()) {
-            // Indices are assigned within column types.
-            std::cmp::Ordering::Equal => self.index.cmp(&other.index),
-            order => order,
-        }
-    }
-}
-
-impl<C: ColumnType> PartialOrd for Column<C> {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl From<ColumnMid> for Column<Any> {
-    fn from(column: ColumnMid) -> Column<Any> {
-        Column {
-            index: column.index,
-            column_type: column.column_type,
-        }
-    }
-}
-
-impl From<Column<Advice>> for Column<Any> {
-    fn from(advice: Column<Advice>) -> Column<Any> {
-        Column {
-            index: advice.index(),
-            column_type: Any::Advice,
-        }
-    }
-}
-
-impl From<Column<Fixed>> for Column<Any> {
-    fn from(advice: Column<Fixed>) -> Column<Any> {
-        Column {
-            index: advice.index(),
-            column_type: Any::Fixed,
-        }
-    }
-}
-
-impl From<Column<Instance>> for Column<Any> {
-    fn from(advice: Column<Instance>) -> Column<Any> {
-        Column {
-            index: advice.index(),
-            column_type: Any::Instance,
-        }
-    }
-}
-
-impl TryFrom<Column<Any>> for Column<Advice> {
-    type Error = &'static str;
-
-    fn try_from(any: Column<Any>) -> Result<Self, Self::Error> {
-        match any.column_type() {
-            Any::Advice => Ok(Column {
-                index: any.index(),
-                column_type: Advice,
-            }),
-            _ => Err("Cannot convert into Column<Advice>"),
-        }
-    }
-}
-
-impl TryFrom<Column<Any>> for Column<Fixed> {
-    type Error = &'static str;
-
-    fn try_from(any: Column<Any>) -> Result<Self, Self::Error> {
-        match any.column_type() {
-            Any::Fixed => Ok(Column {
-                index: any.index(),
-                column_type: Fixed,
-            }),
-            _ => Err("Cannot convert into Column<Fixed>"),
-        }
-    }
-}
-
-impl TryFrom<Column<Any>> for Column<Instance> {
-    type Error = &'static str;
-
-    fn try_from(any: Column<Any>) -> Result<Self, Self::Error> {
-        match any.column_type() {
-            Any::Instance => Ok(Column {
-                index: any.index(),
-                column_type: Instance,
-            }),
-            _ => Err("Cannot convert into Column<Instance>"),
-        }
-    }
-}
-
-pub mod sealed {
-    /// Phase of advice column
-    #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
-    pub struct Phase(pub u8);
-
-    impl Phase {
-        pub fn prev(&self) -> Option<Phase> {
-            self.0.checked_sub(1).map(Phase)
-        }
-    }
-
-    impl SealedPhase for Phase {
-        fn to_sealed(self) -> Phase {
-            self
-        }
-    }
-
-    /// Sealed trait to help keep `Phase` private.
-    pub trait SealedPhase {
-        fn to_sealed(self) -> Phase;
-    }
-}
-
-/// Phase of advice column
-pub trait Phase: SealedPhase {}
-
-impl<P: SealedPhase> Phase for P {}
-
-/// First phase
-#[derive(Debug)]
-pub struct FirstPhase;
-
-impl SealedPhase for FirstPhase {
-    fn to_sealed(self) -> sealed::Phase {
-        sealed::Phase(0)
-    }
-}
-
-/// Second phase
-#[derive(Debug)]
-pub struct SecondPhase;
-
-impl SealedPhase for SecondPhase {
-    fn to_sealed(self) -> sealed::Phase {
-        sealed::Phase(1)
-    }
-}
-
-/// Third phase
-#[derive(Debug)]
-pub struct ThirdPhase;
-
-impl SealedPhase for ThirdPhase {
-    fn to_sealed(self) -> sealed::Phase {
-        sealed::Phase(2)
-    }
-}
-
-/// A selector, representing a fixed boolean value per row of the circuit.
-///
-/// Selectors can be used to conditionally enable (portions of) gates:
-/// ```
-/// use halo2_middleware::poly::Rotation;
-/// # use halo2curves::pasta::Fp;
-/// # use halo2_frontend::plonk::ConstraintSystem;
-///
-/// # let mut meta = ConstraintSystem::<Fp>::default();
-/// let a = meta.advice_column();
-/// let b = meta.advice_column();
-/// let s = meta.selector();
-///
-/// meta.create_gate("foo", |meta| {
-///     let a = meta.query_advice(a, Rotation::prev());
-///     let b = meta.query_advice(b, Rotation::cur());
-///     let s = meta.query_selector(s);
-///
-///     // On rows where the selector is enabled, a is constrained to equal b.
-///     // On rows where the selector is disabled, a and b can take any value.
-///     vec![s * (a - b)]
-/// });
-/// ```
-///
-/// Selectors are disabled on all rows by default, and must be explicitly enabled on each
-/// row when required:
-/// ```
-/// use halo2_frontend::circuit::{Chip, Layouter, Value};
-/// use halo2_frontend::plonk::{Advice, Fixed, Error, Column, Selector};
-/// use halo2_middleware::ff::Field;
-///
-/// struct Config {
-///     a: Column<Advice>,
-///     b: Column<Advice>,
-///     s: Selector,
-/// }
-///
-/// fn circuit_logic<F: Field, C: Chip<F>>(chip: C, mut layouter: impl Layouter<F>) -> Result<(), Error> {
-///     let config = chip.config();
-///     # let config: Config = todo!();
-///     layouter.assign_region(|| "bar", |mut region| {
-///         region.assign_advice(|| "a", config.a, 0, || Value::known(F::ONE))?;
-///         region.assign_advice(|| "a", config.b, 1, || Value::known(F::ONE))?;
-///         config.s.enable(&mut region, 1)
-///     })?;
-///     Ok(())
-/// }
-/// ```
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct Selector(pub usize, pub(crate) bool);
-
-impl Selector {
-    /// Enable this selector at the given offset within the given region.
-    pub fn enable<F: Field>(&self, region: &mut Region<F>, offset: usize) -> Result<(), Error> {
-        region.enable_selector(|| "", self, offset)
-    }
-
-    /// Is this selector "simple"? Simple selectors can only be multiplied
-    /// by expressions that contain no other simple selectors.
-    pub fn is_simple(&self) -> bool {
-        self.1
-    }
-
-    /// Returns index of this selector
-    pub fn index(&self) -> usize {
-        self.0
-    }
-
-    /// Return expression from selector
-    pub fn expr<F: Field>(&self) -> Expression<F> {
-        Expression::Selector(*self)
-    }
-}
-
-/// Query of fixed column at a certain relative location
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct FixedQuery {
-    /// Query index
-    pub index: Option<usize>,
-    /// Column index
-    pub column_index: usize,
-    /// Rotation of this query
-    pub rotation: Rotation,
-}
-
-impl FixedQuery {
-    /// Column index
-    pub fn column_index(&self) -> usize {
-        self.column_index
-    }
-
-    /// Rotation of this query
-    pub fn rotation(&self) -> Rotation {
-        self.rotation
-    }
-}
-
-/// Query of advice column at a certain relative location
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct AdviceQuery {
-    /// Query index
-    pub index: Option<usize>,
-    /// Column index
-    pub column_index: usize,
-    /// Rotation of this query
-    pub rotation: Rotation,
-}
-
-impl AdviceQuery {
-    /// Column index
-    pub fn column_index(&self) -> usize {
-        self.column_index
-    }
-
-    /// Rotation of this query
-    pub fn rotation(&self) -> Rotation {
-        self.rotation
-    }
-}
-
-/// Query of instance column at a certain relative location
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct InstanceQuery {
-    /// Query index
-    pub index: Option<usize>,
-    /// Column index
-    pub column_index: usize,
-    /// Rotation of this query
-    pub rotation: Rotation,
-}
-
-impl InstanceQuery {
-    /// Column index
-    pub fn column_index(&self) -> usize {
-        self.column_index
-    }
-
-    /// Rotation of this query
-    pub fn rotation(&self) -> Rotation {
-        self.rotation
-    }
-}
-
-/// A fixed column of a lookup table.
-///
-/// A lookup table can be loaded into this column via [`Layouter::assign_table`]. Columns
-/// can currently only contain a single table, but they may be used in multiple lookup
-/// arguments via [`super::constraint_system::ConstraintSystem::lookup`].
-///
-/// Lookup table columns are always "encumbered" by the lookup arguments they are used in;
-/// they cannot simultaneously be used as general fixed columns.
-///
-/// [`Layouter::assign_table`]: crate::circuit::Layouter::assign_table
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
-pub struct TableColumn {
-    /// The fixed column that this table column is stored in.
-    ///
-    /// # Security
-    ///
-    /// This inner column MUST NOT be exposed in the public API, or else chip developers
-    /// can load lookup tables into their circuits without default-value-filling the
-    /// columns, which can cause soundness bugs.
-    pub(super) inner: Column<Fixed>,
-}
-
-impl TableColumn {
-    /// Returns inner column
-    pub fn inner(&self) -> Column<Fixed> {
-        self.inner
-    }
-}
-
-/// A challenge squeezed from transcript after advice columns at the phase have been committed.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub struct Challenge {
-    pub index: usize,
-    pub(crate) phase: u8,
-}
-
-impl Challenge {
-    /// Index of this challenge.
-    pub fn index(&self) -> usize {
-        self.index
-    }
-
-    /// Phase of this challenge.
-    pub fn phase(&self) -> u8 {
-        self.phase
-    }
-
-    /// Return Expression
-    pub fn expr<F: Field>(&self) -> Expression<F> {
-        Expression::Challenge(*self)
-    }
-}
-
-impl From<Challenge> for ChallengeMid {
-    fn from(val: Challenge) -> Self {
-        ChallengeMid {
-            index: val.index,
-            phase: val.phase,
-        }
-    }
-}
-
-impl From<ChallengeMid> for Challenge {
-    fn from(c: ChallengeMid) -> Self {
-        Self {
-            index: c.index,
-            phase: c.phase,
-        }
-    }
-}
+use std::ops::{Neg, Sub};
 
 /// Low-degree expression representing an identity that must hold over the committed columns.
 #[derive(Clone, PartialEq, Eq)]
